backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Speech Recognition
openai-whisper==20231117
speechrecognition==3.10.1
pyaudio==0.2.14
webrtcvad==2.0.10
numpy==1.26.3

# Text-to-Speech
pyttsx3==2.90
edge-tts==6.1.9
pygame==2.5.2

# Local LLM
ollama==0.1.7
requests==2.31.0
aiohttp==3.9.1

# System Control
pyautogui==0.9.54
pygetwindow==0.0.9
psutil==5.9.8
pycaw==20230407
comtypes==1.2.1
pywin32==306; platform_system=="Windows"

# macOS specific
pyobjc-framework-Quartz==10.1; platform_system=="Darwin"
pyobjc-framework-ApplicationServices==10.1; platform_system=="Darwin"
osascript==2020.12.3; platform_system=="Darwin"

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
asyncio-mqtt==0.16.1
